adsd

+++ andmed.php
@@ -27,7 +27,7 @@ if($result->num_rows != 0){
     while($rows = $result->fetch_assoc()){
             $abilist = array();
             $Nimi = $rows["Nimi"];
-            $Hääli = intval($rows["Hääli"]);
+            $Hääli = (double)intval($rows["Hääli"]);
             array_push($abilist, $Nimi, $Hääli);
             array_push($valjalist, $abilist);
